+++ crates/rattler_repodata_gateway/src/gateway/sharded_subdir/mod.rs
@@ -1,17 +1,20 @@
 use std::{borrow::Cow, io::Write, path::PathBuf, sync::Arc};
 
-use http::{header::CACHE_CONTROL, HeaderValue, StatusCode};
-use rattler_conda_types::{Channel, PackageName, RepoDataRecord, Shard, ShardedRepodata};
-use reqwest_middleware::ClientWithMiddleware;
-use simple_spawn_blocking::tokio::run_blocking_task;
-use url::Url;
-
 use crate::{
     fetch::{CacheAction, FetchRepoDataError},
     gateway::{error::SubdirNotFoundError, subdir::SubdirClient},
     reporter::ResponseReporterExt,
     GatewayError, Reporter,
 };
+use fs_err::tokio as tokio_fs;
+use http::{header::CACHE_CONTROL, HeaderValue, StatusCode};
+use rattler_conda_types::{
+    Channel, ChannelUrl, PackageName, RepoDataRecord, Shard, ShardedRepodata,
+};
+use rattler_redaction::Redact;
+use reqwest_middleware::ClientWithMiddleware;
+use simple_spawn_blocking::tokio::run_blocking_task;
+use url::Url;
 
 mod index;
 
@@ -86,7 +89,7 @@ impl ShardedSubdir {
 
         // Determine the cache directory and make sure it exists.
         let cache_dir = cache_dir.join("shards-v1");
-        tokio::fs::create_dir_all(&cache_dir)
+        tokio_fs::create_dir_all(&cache_dir)
             .await
             .map_err(FetchRepoDataError::IoError)?;
 
@@ -120,12 +123,12 @@ impl SubdirClient for ShardedSubdir {
 
         // Read the cached shard
         if self.cache_action != CacheAction::NoCache {
-            match tokio::fs::read(&shard_cache_path).await {
+            match tokio_fs::read(&shard_cache_path).await {
                 Ok(cached_bytes) => {
                     // Decode the cached shard
                     return parse_records(
                         cached_bytes,
-                        self.channel.canonical_name(),
+                        self.channel.base_url.clone(),
                         self.package_base_url.clone(),
                     )
                     .await
@@ -192,7 +195,7 @@ impl SubdirClient for ShardedSubdir {
         // Create a future to parse the records from the shard
         let parse_records_fut = parse_records(
             shard_bytes,
-            self.channel.canonical_name(),
+            self.channel.base_url.clone(),
             self.package_base_url.clone(),
         );
 
@@ -273,7 +276,7 @@ async fn decode_zst_bytes_async<R: AsRef<[u8]> + Send + 'static>(
 
 async fn parse_records<R: AsRef<[u8]> + Send + 'static>(
     bytes: R,
-    channel_name: String,
+    channel_base_url: ChannelUrl,
     base_url: Url,
 ) -> Result<Vec<RepoDataRecord>, GatewayError> {
     run_blocking_task(move || {
@@ -291,7 +294,7 @@ async fn parse_records<R: AsRef<[u8]> + Send + 'static>(
                 url: base_url
                     .join(&file_name)
                     .expect("filename is not a valid url"),
-                channel: channel_name.clone(),
+                channel: Some(channel_base_url.url().clone().redact().to_string()),
                 package_record,
                 file_name,
             })