+++ py-rattler/rattler/networking/client.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 from rattler.rattler import PyClientWithMiddleware
-from rattler.networking.middleware import AuthenticationMiddleware, MirrorMiddleware
+from rattler.networking.middleware import AuthenticationMiddleware, MirrorMiddleware, OciMiddleware, GCSMiddleware
 
 
 class Client:
@@ -8,7 +8,10 @@ class Client:
     A client that can be used to make requests.
     """
 
-    def __init__(self, middlewares: list[AuthenticationMiddleware | MirrorMiddleware] | None = None) -> None:
+    def __init__(
+        self,
+        middlewares: list[AuthenticationMiddleware | MirrorMiddleware | OciMiddleware | GCSMiddleware] | None = None,
+    ) -> None:
         self._client = PyClientWithMiddleware(
             [middleware._middleware for middleware in middlewares] if middlewares else None
         )