+++ py-rattler/rattler/networking/middleware.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
-from rattler.rattler import PyMirrorMiddleware, PyAuthenticationMiddleware
+from rattler.rattler import PyMirrorMiddleware, PyAuthenticationMiddleware, PyOciMiddleware, PyGCSMiddleware
 
 
 class MirrorMiddleware:
@@ -41,6 +41,10 @@ class MirrorMiddleware:
 
 
 class AuthenticationMiddleware:
+    """
+    Middleware to handle authentication from keychain
+    """
+
     def __init__(self) -> None:
         self._middleware = PyAuthenticationMiddleware()
 
@@ -60,4 +64,53 @@ class AuthenticationMiddleware:
         >>>
         ```
         """
+        return f"{type(self).__name__}()"
+
+
+class OciMiddleware:
+    """
+    Middleware to handle `oci://` URLs
+    """
+
+    def __init__(self) -> None:
+        self._middleware = PyOciMiddleware()
+
+    def __repr__(self) -> str:
+        """
+        Returns a representation of the Middleware
+
+        Examples
+        --------
+        ```python
+        >>> from rattler.networking import Client
+        >>> middleware = OciMiddleware()
+        >>> middleware
+        OciMiddleware()
+        >>> Client([middleware])
+        Client()
+        >>>
+        ```
+        """
+        return f"{type(self).__name__}()"
+
+
+class GCSMiddleware:
+    """
+    Middleware to work with gcs:// URLs
+
+    Examples
+    --------
+    ```python
+    >>> from rattler.networking import Client
+    >>> middleware = GCSMiddleware()
+    >>> middleware
+    GCSMiddleware()
+    >>> Client([middleware])
+    Client()
+    """
+
+    def __init__(self) -> None:
+        self._middleware = PyGCSMiddleware()
+
+    def __repr__(self) -> str:
         return f"{type(self).__name__}()"