+++ py-rattler/src/networking/client.rs
@@ -1,6 +1,8 @@
 use crate::networking::middleware::PyMiddleware;
 use pyo3::{pyclass, pymethods};
-use rattler_networking::{AuthenticationMiddleware, MirrorMiddleware};
+use rattler_networking::{
+    AuthenticationMiddleware, GCSMiddleware, MirrorMiddleware, OciMiddleware,
+};
 use reqwest_middleware::ClientWithMiddleware;
 
 #[pyclass]
@@ -20,12 +22,12 @@ impl PyClientWithMiddleware {
         let client = middlewares
             .into_iter()
             .fold(client, |client, middleware| match middleware {
-                PyMiddleware::MirrorMiddleware(middleware) => {
-                    client.with(MirrorMiddleware::from(middleware))
-                }
-                PyMiddleware::AuthenticationMiddleware(middleware) => {
+                PyMiddleware::Mirror(middleware) => client.with(MirrorMiddleware::from(middleware)),
+                PyMiddleware::Authentication(middleware) => {
                     client.with(AuthenticationMiddleware::from(middleware))
                 }
+                PyMiddleware::Oci(middleware) => client.with(OciMiddleware::from(middleware)),
+                PyMiddleware::Gcs(middleware) => client.with(GCSMiddleware::from(middleware)),
             });
         let client = client.build();