+++ crates/rattler_cache/CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.11](https://github.com/conda/rattler/compare/rattler_cache-v0.2.10...rattler_cache-v0.2.11) - 2024-11-18
+
+### Other
+
+- updated the following local packages: rattler_networking
+
+## [0.2.10](https://github.com/conda/rattler/compare/rattler_cache-v0.2.9...rattler_cache-v0.2.10) - 2024-11-18
+
+### Other
+
+- updated the following local packages: rattler_conda_types
+
 ## [0.2.9](https://github.com/conda/rattler/compare/rattler_cache-v0.2.8...rattler_cache-v0.2.9) - 2024-11-14
 
 ### Added